made changes in migration

+++ database/migrations/2017_10_31_135945_create_customers_table.php
@@ -28,7 +28,7 @@ class CreateCustomersTable extends Migration
                 Gender enum('M', 'F', 'O') NOT NULL,
                 Date_of_birth date NOT NULL,
                 Email varchar(25) NOT NULL UNIQUE,
-                Aadhar_number int NOT NULL UNIQUE,
+                Aadhar_number bigint NOT NULL UNIQUE,
                 Account_Number bigint NOT NULL
             )
         ");
