+++ database/migrations/2017_10_31_141208_create_employees_table.php
@@ -26,7 +26,7 @@ class CreateEmployeesTable extends Migration
                 Gender enum('M','F','O') NOT NULL,
                 Date_Of_birth date NOT NULL,
                 Email varchar(25) NOT NULL UNIQUE,
-                Aadhar_number int NOT NULL UNIQUE,
+                Aadhar_number bigint NOT NULL UNIQUE,
                 Designation text NOT NULL,
                 Salary int NOT NULL,
                 IFSC_Code varchar(15) NOT NULL,
